--- client/src/Header.css
@@ -1,27 +1,0 @@
-.login-btn {
-    float: right;
-}
-
-.logout-btn {
-    float: right;
-}
-
-.header-btn {
-    float: right;
-}
-
-.categoryTab, .lawyerTab {
-    float: right;
-    color: white;
-    padding-top: 15px;
-    padding-right: 15px;
-    padding-left: 5px;
-}
-
-h3:hover {
-    color: black;
-}
-
-.request-btn {
-    float: right;
-}
